Internal overlay
- Add overlay to internal vertex format.
- Move interval vertex format related constants to helper class.
- Switch strides to long so java stops complaining about implicit casts
  in multiplication.
- Remove VertexTypes#init as it's not needed anymore.

+++ src/main/java/com/jozufozu/flywheel/lib/vertex/BlockVertexList.java
@@ -8,7 +8,7 @@ import com.jozufozu.flywheel.lib.math.RenderMath;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 
 public class BlockVertexList extends AbstractVertexList {
-	private static final int STRIDE = 32;
+	private static final long STRIDE = 32;
 
 	@Override
 	public float x(int index) {
